+++ Sources/BugsnagPerformance/Private/BugsnagPerformanceLibrary.h
@@ -25,8 +25,6 @@ public:
     static std::shared_ptr<BugsnagPerformanceImpl> getBugsnagPerformanceImpl() noexcept;
     static std::shared_ptr<Reachability> getReachability() noexcept;
     static AppStateTracker *getAppStateTracker() noexcept;
-
-    static void testing_reset();
 private:
     // Use GNU constructor attribute to auto-call functions before main() is called.
     // https://gcc.gnu.org/onlinedocs/gcc-4.7.0/gcc/Function-Attributes.html