+++ Sources/BugsnagPerformance/Private/BugsnagPerformanceLibrary.mm
@@ -24,27 +24,57 @@ BugsnagPerformanceLibrary &BugsnagPerformanceLibrary::sharedInstance() noexcept 
 }
 
 void BugsnagPerformanceLibrary::calledAsEarlyAsPossible() noexcept {
-    auto instance = sharedInstance();
-    auto config = [BSGEarlyConfiguration new];
-    instance.earlyConfigure(config);
-    instance.earlySetup();
+    @autoreleasepool {
+        BSGLogDebug(@"BugsnagPerformanceLibrary::calledAsEarlyAsPossible");
+        auto instance = sharedInstance();
+        auto config = [BSGEarlyConfiguration new];
+        instance.earlyConfigure(config);
+        instance.earlySetup();
+    }
 }
 
 void BugsnagPerformanceLibrary::calledRightBeforeMain() noexcept {
-    sharedInstance().bugsnagPerformanceImpl_->willCallMainFunction();
+    @autoreleasepool {
+        BSGLogDebug(@"BugsnagPerformanceLibrary::calledRightBeforeMain");
+        sharedInstance().bugsnagPerformanceImpl_->willCallMainFunction();
+    }
 }
 
 void BugsnagPerformanceLibrary::configureLibrary(BugsnagPerformanceConfiguration *config) noexcept {
-    NSError *__autoreleasing error = nil;
-    if (![config validate:&error]) {
-        BSGLogError(@"Configuration validation failed with error: %@", error);
-    }
+    @autoreleasepool {
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: apiKey = %@", config.apiKey);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: endpoint = %@", config.endpoint);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: autoInstrumentAppStarts = %d", config.autoInstrumentAppStarts);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: autoInstrumentViewControllers = %d", config.autoInstrumentViewControllers);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: autoInstrumentNetworkRequests = %d", config.autoInstrumentNetworkRequests);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: appVersion = %@", config.appVersion);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: bundleVersion = %@", config.bundleVersion);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: releaseStage = %@", config.releaseStage);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: enabledReleaseStages = %@", config.enabledReleaseStages);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: tracePropagationUrls = %@", config.tracePropagationUrls);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: clearPersistenceOnStart = %d", config.internal.clearPersistenceOnStart);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: autoTriggerExportOnBatchSize = %llu", config.internal.autoTriggerExportOnBatchSize);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: performWorkInterval = %f", config.internal.performWorkInterval);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: maxRetryAge = %f", config.internal.maxRetryAge);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: probabilityValueExpiresAfterSeconds = %f", config.internal.probabilityValueExpiresAfterSeconds);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: probabilityRequestsPauseForSeconds = %f", config.internal.probabilityRequestsPauseForSeconds);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: initialSamplingProbability = %f", config.internal.initialSamplingProbability);
+        BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary: maxPackageContentLength = %llu", config.internal.maxPackageContentLength);
 
-    sharedInstance().configure(config);
+        NSError *__autoreleasing error = nil;
+        if (![config validate:&error]) {
+            BSGLogError(@"Configuration validation failed with error: %@", error);
+        }
+
+        sharedInstance().configure(config);
+    }
 }
 
 void BugsnagPerformanceLibrary::startLibrary() noexcept {
-    sharedInstance().start();
+    @autoreleasepool {
+        BSGLogDebug(@"BugsnagPerformanceLibrary::startLibrary");
+        sharedInstance().start();
+    }
 }
 
 BugsnagPerformanceLibrary::BugsnagPerformanceLibrary()
@@ -54,10 +84,12 @@ BugsnagPerformanceLibrary::BugsnagPerformanceLibrary()
 {}
 
 void BugsnagPerformanceLibrary::earlyConfigure(BSGEarlyConfiguration *config) noexcept {
+    BSGLogDebug(@"BugsnagPerformanceLibrary::configureLibrary");
     bugsnagPerformanceImpl_->earlyConfigure(config);
 }
 
 void BugsnagPerformanceLibrary::earlySetup() noexcept {
+    BSGLogDebug(@"BugsnagPerformanceLibrary::earlySetup");
     auto impl = bugsnagPerformanceImpl_;
     bugsnagPerformanceImpl_->setOnViewLoadSpanStarted([=](NSString *className) {
         impl->didStartViewLoadSpan(className);
@@ -66,10 +98,12 @@ void BugsnagPerformanceLibrary::earlySetup() noexcept {
 }
 
 void BugsnagPerformanceLibrary::configure(BugsnagPerformanceConfiguration *config) noexcept {
+    BSGLogDebug(@"BugsnagPerformanceLibrary::configure");
     bugsnagPerformanceImpl_->configure(config);
 }
 
 void BugsnagPerformanceLibrary::start() noexcept {
+    BSGLogDebug(@"BugsnagPerformanceLibrary::start");
     bugsnagPerformanceImpl_->start();
 }
 
@@ -83,16 +117,4 @@ std::shared_ptr<Reachability> BugsnagPerformanceLibrary::getReachability() noexc
 
 AppStateTracker *BugsnagPerformanceLibrary::getAppStateTracker() noexcept {
     return sharedInstance().appStateTracker_;
-}
-
-// Keep old instances around while testing so that lingering callbacks don't reference
-// a defunct instance.
-[[clang::no_destroy]]
-static std::vector<std::shared_ptr<BugsnagPerformanceLibrary>> testing_previous_instances;
-
-void BugsnagPerformanceLibrary::testing_reset() {
-    testing_previous_instances.push_back(instance_do_not_access_directly);
-    instance_do_not_access_directly.reset();
-    calledAsEarlyAsPossible();
-    calledRightBeforeMain();
 }