+++ features/fixtures/ios/Fixture/ViewController.swift
@@ -9,6 +9,7 @@ import UIKit
 
 class ViewController: UIViewController {
     var fixture: Fixture = Fixture()
+//    var fixture: Fixture = PresetFixture(scenarioName: "AutoInstrumentAVAssetScenario")
     
     required init?(coder: NSCoder) {
         super.init(coder: coder)