+++ features/steps/app_steps.rb
@@ -20,6 +20,10 @@ Before('@skip_ios_15_and_above') do |_scenario|
   skip_above('ios', 14.99)
 end
 
+Before('@skip') do |_scenario|
+  skip_this_scenario("Skipping scenario")
+end
+
 Then('I discard every {request_type}') do |request_type|
   until Maze::Server.list_for(request_type).current.nil?
     Maze::Server.list_for(request_type).next
@@ -167,6 +171,54 @@ Then('a span integer attribute {string} is greater than {int}') do |attribute, e
   selected_attributes = spans.map { |span| span['attributes'].find { |a| a['key'].eql?(attribute) && a['value'].has_key?('intValue') } }.compact
   attribute_values = selected_attributes.map { |a| a['value']['intValue'].to_i > expected }
   Maze.check.false(attribute_values.empty?)
+end
+
+Then('a span array attribute {string} contains the string value {string} at index {int}') do |attribute, expected, index|
+  value = get_array_value_at_index(attribute, index, 'stringValue')
+  Maze.check.true(value == expected)
+end
+
+Then('a span array attribute {string} contains the integer value {int} at index {int}') do |attribute, expected, index|
+  value = get_array_value_at_index(attribute, index, 'intValue')
+  Maze.check.true(value.to_i == expected)
+end
+
+Then('a span array attribute {string} contains the float value {float} at index {int}') do |attribute, expected, index|
+  value = get_array_value_at_index(attribute, index, 'doubleValue')
+  Maze.check.true(value == expected)
+end
+
+Then('a span array attribute {string} contains the value true at index {int}') do |attribute, index|
+  value = get_array_value_at_index(attribute, index, 'boolValue')
+  Maze.check.true(value == true)
+end
+
+Then('a span array attribute {string} contains the value false at index {int}') do |attribute, index|
+  value = get_array_value_at_index(attribute, index, 'boolValue')
+  Maze.check.true(value == false)
+end
+
+def get_array_value_at_index(attribute, index, type)
+  array = get_array_attribute_contents(attribute)
+  Maze.check.true(array.length() > index)
+  value = array[index]
+  Maze.check.true(value.has_key?(type))
+  return value[type]
+end
+
+def get_array_attribute_contents(attribute)
+  spans = spans_from_request_list(Maze::Server.list_for('traces'))
+  selected_attributes = spans.map { |span| span['attributes'].find { |a| a['key'].eql?(attribute) &&
+                                                                         a['value'].has_key?('arrayValue') &&
+                                                                         a['value']['arrayValue'].has_key?('values') } }.compact
+  array_attributes = selected_attributes.map { |a| a['value']['arrayValue']['values'] }
+  Maze.check.false(array_attributes.empty?)
+  return array_attributes[0]
+end
+
+Then('a span array attribute {string} is empty') do |attribute|
+  array_contents = get_array_attribute_contents(attribute)
+  Maze.check.true(array_contents.empty?)
 end
 
 Then('a span named {string} is a child of span named {string}') do |name1, name2|